[color-picker] deprecated Item element for colors
